+++ Integrations/MISP_V2/test_data/test_constants.py
@@ -1,0 +1,328 @@
+full_response_before_filtering = [
+    {
+        "Event": {
+            "Attribute": [
+                {
+                    "Galaxy": [],
+                    "ShadowAttribute": [],
+                    "category": "Payload installation",
+                    "comment": "mock",
+                    "deleted": False,
+                    "disable_correlation": False,
+                    "distribution": "5",
+                    "event_id": "2193",
+                    "id": "128241",
+                    "object_id": "0",
+                    "object_relation": None,
+                    "sharing_group_id": "0",
+                    "timestamp": "1461142838",
+                    "to_ids": False,
+                    "type": "md5",
+                    "uuid": "57174536-95fc-474a-bb9f-46dc950d210f",
+                    "value": "mock"
+                },
+                {
+                    "Galaxy": [],
+                    "ShadowAttribute": [],
+                    "category": "Payload installation",
+                    "comment": "mock",
+                    "deleted": False,
+                    "disable_correlation": False,
+                    "distribution": "5",
+                    "event_id": "2193",
+                    "id": "128242",
+                    "object_id": "0",
+                    "object_relation": None,
+                    "sharing_group_id": "0",
+                    "timestamp": "1461145369",
+                    "to_ids": False,
+                    "type": "sha256",
+                    "uuid": "57174f19-f06c-4809-92b3-476d02de0b81",
+                    "value": "mock"},
+                {
+                    "Galaxy": [],
+                    "ShadowAttribute": [],
+                    "category": "Payload installation",
+                    "comment": "mock",
+                    "deleted": False,
+                    "disable_correlation": False,
+                    "distribution": "5",
+                    "event_id": "2193",
+                    "id": "128243",
+                    "object_id": "0",
+                    "object_relation": None,
+                    "sharing_group_id": "0",
+                    "timestamp": "1461145369",
+                    "to_ids": False,
+                    "type": "sha1",
+                    "uuid": "57174f19-c3d8-43f0-96bb-43e402de0b81",
+                    "value": "mock"},
+                {
+                    "Galaxy": [],
+                    "ShadowAttribute": [],
+                    "category": "External analysis",
+                    "comment": "mock",
+                    "deleted": False,
+                    "disable_correlation": False,
+                    "distribution": "5",
+                    "event_id": "2193",
+                    "id": "128244",
+                    "object_id": "0",
+                    "object_relation": None,
+                    "sharing_group_id": "0",
+                    "timestamp": "1461145370",
+                    "to_ids": False,
+                    "type": "link",
+                    "uuid": "57174f1a-a47c-458d-9900-451e02de0b81",
+                    "value": "mock"},
+                {
+                    "Galaxy": [],
+                    "ShadowAttribute": [],
+                    "category": "Network activity",
+                    "comment": "mock",
+                    "deleted": False,
+                    "disable_correlation": False,
+                    "distribution": "5",
+                    "event_id": "2193",
+                    "id": "128245",
+                    "object_id": "0",
+                    "object_relation": None,
+                    "sharing_group_id": "0",
+                    "timestamp": "1461152759",
+                    "to_ids": False,
+                    "type": "hostname",
+                    "uuid": "57176bf7-4e28-4e60-88d4-4dea950d210f",
+                    "value": "mock"
+                },
+                {
+                    "Galaxy": [],
+                    "ShadowAttribute": [],
+                    "category": "Network activity",
+                    "comment": "mock",
+                    "deleted": False,
+                    "disable_correlation": False,
+                    "distribution": "5",
+                    "event_id": "2193",
+                    "id": "128246",
+                    "object_id": "0",
+                    "object_relation": None,
+                    "sharing_group_id": "0",
+                    "timestamp": "1461152810",
+                    "to_ids": False,
+                    "type": "url",
+                    "uuid": "57176c2a-2280-4f00-9208-42c1950d210f",
+                    "value": "mock"
+                },
+                {
+                    "Galaxy": [],
+                    "ShadowAttribute": [],
+                    "category": "Network activity",
+                    "comment": "mock",
+                    "deleted": False,
+                    "disable_correlation": False,
+                    "distribution": "5",
+                    "event_id": "2193",
+                    "id": "128247",
+                    "object_id": "0",
+                    "object_relation": None,
+                    "sharing_group_id": "0",
+                    "timestamp": "1461152872",
+                    "to_ids": False,
+                    "type": "ip-dst",
+                    "uuid": "57176c68-3b48-47d1-a549-4f20950d210f",
+                    "value": "mock"
+                },
+                {
+                    "Galaxy": [],
+                    "ShadowAttribute": [],
+                    "category": "Network activity",
+                    "comment": "mock",
+                    "deleted": False,
+                    "disable_correlation": False,
+                    "distribution": "5",
+                    "event_id": "2193",
+                    "id": "128248",
+                    "object_id": "0",
+                    "object_relation": None,
+                    "sharing_group_id": "0",
+                    "timestamp": "1461152936",
+                    "to_ids": False,
+                    "type": "url",
+                    "uuid": "57176ca8-4fd8-4eac-a432-4fdb950d210f",
+                    "value": "mock"
+                },
+                {
+                    "Galaxy": [],
+                    "ShadowAttribute": [],
+                    "category": "Other",
+                    "comment": "mock",
+                    "deleted": False,
+                    "disable_correlation": False,
+                    "distribution": "5",
+                    "event_id": "2193",
+                    "id": "128249",
+                    "object_id": "0",
+                    "object_relation": None,
+                    "sharing_group_id": "0",
+                    "timestamp": "1461153063",
+                    "to_ids": False,
+                    "type": "comment",
+                    "uuid": "57176d27-f044-4038-9604-49da950d210f",
+                    "value": "mock"
+                },
+                {
+                    "Galaxy": [],
+                    "ShadowAttribute": [],
+                    "category": "Other",
+                    "comment": "mock",
+                    "deleted": False,
+                    "disable_correlation": False,
+                    "distribution": "5",
+                    "event_id": "2193",
+                    "id": "128250",
+                    "object_id": "0",
+                    "object_relation": None,
+                    "sharing_group_id": "0",
+                    "timestamp": "1461153134",
+                    "to_ids": False,
+                    "type": "comment",
+                    "uuid": "57176d6e-a4e8-4c08-bd9c-41d9950d210f",
+                    "value": "mock"
+                }
+            ],
+            "Galaxy": [],
+            "Object": [],
+            "Org": {
+                "id": "1",
+                "name": "ORGNAME",
+                "uuid": "5c77af86-8274-4106-862e-0045ac110002"
+            },
+            "Orgc": {
+                "id": "2",
+                "name": "CIRCL",
+                "uuid": "55f6ea5e-2c60-40e5-964f-47a8950d210f"
+            },
+            "RelatedEvent": [],
+            "ShadowAttribute": [],
+            "Tag": [
+                {
+                    "colour": "#ffffff",
+                    "exportable": True,
+                    "hide_tag": False,
+                    "id": "2",
+                    "name": "tlp:white",
+                    "numerical_value": None,
+                    "user_id": "0"
+                }
+            ],
+            "analysis": "0",
+            "attribute_count": "10",
+            "date": "2016-04-20",
+            "disable_correlation": False,
+            "distribution": "1",
+            "extends_uuid": "",
+            "id": "2193",
+            "info": "New ransomware",
+            "locked": False,
+            "org_id": "1",
+            "orgc_id": "2",
+            "proposal_email_lock": False,
+            "publish_timestamp": "1563282468",
+            "published": True,
+            "sharing_group_id": "0",
+            "threat_level_id": "3",
+            "timestamp": "1461153134",
+            "uuid": "57174526-23d8-4895-8c08-4ed1950d210f"
+        }
+    }
+]
+
+response_after_filtering_category_eventid_uuid = [
+    {
+        "Analysis": "0",
+        "Attribute": [
+            {
+                "Category": "Payload installation",
+                "EventID": "2193",
+                "UUID": "57174536-95fc-474a-bb9f-46dc950d210f"
+            },
+            {
+                "Category": "Payload installation",
+                "EventID": "2193",
+                "UUID": "57174f19-f06c-4809-92b3-476d02de0b81"
+            },
+            {
+                "Category": "Payload installation",
+                "EventID": "2193",
+                "UUID": "57174f19-c3d8-43f0-96bb-43e402de0b81"
+            },
+            {
+                "Category": "External analysis",
+                "EventID": "2193",
+                "UUID": "57174f1a-a47c-458d-9900-451e02de0b81"
+            },
+            {
+                "Category": "Network activity",
+                "EventID": "2193",
+                "UUID": "57176bf7-4e28-4e60-88d4-4dea950d210f"
+            },
+            {
+                "Category": "Network activity",
+                "EventID": "2193",
+                "UUID": "57176c2a-2280-4f00-9208-42c1950d210f"
+            },
+            {
+                "Category": "Network activity",
+                "EventID": "2193",
+                "UUID": "57176c68-3b48-47d1-a549-4f20950d210f"
+            },
+            {
+                "Category": "Network activity",
+                "EventID": "2193",
+                "UUID": "57176ca8-4fd8-4eac-a432-4fdb950d210f"
+            },
+            {
+                "Category": "Other",
+                "EventID": "2193",
+                "UUID": "57176d27-f044-4038-9604-49da950d210f"
+            },
+            {
+                "Category": "Other",
+                "EventID": "2193",
+                "UUID": "57176d6e-a4e8-4c08-bd9c-41d9950d210f"
+            }
+        ],
+        "Date": "2016-04-20",
+        "DisableCorrelation": False,
+        "Distribution": "1",
+        "Galaxy": [],
+        "ID": "2193",
+        "Info": "New ransomware",
+        "Locked": False,
+        "Object": [],
+        "Organisation": {
+            "ID": "1",
+            "Name": "ORGNAME",
+            "UUID": "5c77af86-8274-4106-862e-0045ac110002"
+        },
+        "OwnerOrganisation": {
+            "ID": "2",
+            "Name": "CIRCL",
+            "UUID": "55f6ea5e-2c60-40e5-964f-47a8950d210f"
+        },
+        "ProposalEmailLock": False,
+        "PublishTimestamp": "1563282468",
+        "Published": True,
+        "RelatedEvent": [],
+        "ShadowAttribute": [],
+        "SharingGroupID": "0",
+        "Tag": [
+            {
+                "Name": "tlp:white"
+            }
+        ],
+        "ThreatLevelID": "3",
+        "Timestamp": "1461153134",
+        "UUID": "57174526-23d8-4895-8c08-4ed1950d210f"
+    }
+]