+++ Scripts/CommonServerPython/CommonServerPython_test.py
@@ -10,7 +10,8 @@ import pytest
 from CommonServerPython import xml2json, json2xml, entryTypes, formats, tableToMarkdown, underscoreToCamelCase, \
     flattenCell, date_to_timestamp, datetime, camelize, pascalToSpace, argToList, \
     remove_nulls_from_dictionary, is_error, get_error, hash_djb2, fileResult, is_ip_valid, get_demisto_version, \
-    IntegrationLogger, parse_date_string, IS_PY3, DebugLogger, b64_encode, parse_date_range, return_outputs
+    IntegrationLogger, parse_date_string, IS_PY3, DebugLogger, b64_encode, parse_date_range, return_outputs, \
+    argToBoolean, batch
 
 try:
     from StringIO import StringIO
@@ -819,6 +820,15 @@ class TestBaseClient:
         with raises(DemistoException, match="Error in API call"):
             self.client._http_request('get', 'event')
 
+    def test_http_request_not_ok_with_json_parsing(self, requests_mock):
+        from CommonServerPython import DemistoException
+        requests_mock.get('http://example.com/api/v2/event', status_code=500, content=str.encode(json.dumps(self.text)))
+        with raises(DemistoException) as exception:
+            self.client._http_request('get', 'event')
+        message = str(exception.value)
+        response_json_error = json.loads(message.split('\n')[1])
+        assert response_json_error == self.text
+
     def test_http_request_timeout(self, requests_mock):
         from CommonServerPython import DemistoException
         requests_mock.get('http://example.com/api/v2/event', exc=requests.exceptions.ConnectTimeout)
@@ -986,3 +996,28 @@ class TestReturnOutputs:
         assert outputs == results['Contents']
         assert outputs == results['EntryContext']
         assert md == results['HumanReadable']
+
+
+def test_argToBoolean():
+    assert argToBoolean('true') is True
+    assert argToBoolean('yes') is True
+    assert argToBoolean('TrUe') is True
+    assert argToBoolean(True) is True
+
+    assert argToBoolean('false') is False
+    assert argToBoolean('no') is False
+    assert argToBoolean(False) is False
+
+
+batch_params = [
+    # full batch
+    ([1, 2, 3], 2, [[1, 2], [3]]),
+    ([], 1, []),
+    ([1, 2, 3], 5, [[1, 2, 3]])
+]
+
+
+@pytest.mark.parametrize('iterable, sz, expected', batch_params)
+def test_batch(iterable, sz, expected):
+    for i, item in enumerate(batch(iterable, sz)):
+        assert expected[i] == item