+++ Tests/scripts/slack_notifier.py
@@ -8,22 +8,19 @@ from slackclient import SlackClient
 
 from Tests.test_utils import str2bool, run_command, LOG_COLORS, print_color
 
-DEMISTO_GREY_ICON = 'https://3xqz5p387rui1hjtdv1up7lw-wpengine.netdna-ssl.com/wp-content/uploads/2018/07/Demisto-Icon-Dark.png'
+DEMISTO_GREY_ICON = 'https://3xqz5p387rui1hjtdv1up7lw-wpengine.netdna-ssl.com/wp-content/' \
+                    'uploads/2018/07/Demisto-Icon-Dark.png'
 
 
 def http_request(url, params_dict=None):
-    try:
-        res = requests.request("GET",
-                               url,
-                               verify=True,
-                               params=params_dict,
-                               )
-        res.raise_for_status()
+    res = requests.request("GET",
+                           url,
+                           verify=True,
+                           params=params_dict,
+                           )
+    res.raise_for_status()
 
-        return res.json()
-
-    except Exception, e:
-        raise e
+    return res.json()
 
 
 def options_handler():
@@ -33,27 +30,32 @@ def options_handler():
     parser.add_argument('-b', '--buildNumber', help='The build number', required=True)
     parser.add_argument('-s', '--slack', help='The token for slack', required=True)
     parser.add_argument('-c', '--circleci', help='The token for circleci', required=True)
-    parser.add_argument('-f', '--env_results_file_name', help='The env results file containing the dns address', required=True)
+    parser.add_argument('-f', '--env_results_file_name', help='The env results file containing the dns address',
+                        required=True)
     options = parser.parse_args()
 
     return options
 
 
 def get_attachments(build_url, env_results_file_name):
-    content_team_fields, content_fields, failed_tests = get_fields()
-    color = 'good' if not failed_tests else 'danger'
-    title = 'Content Build - Success' if not failed_tests else 'Content Build - Failure'
-
     with open(env_results_file_name, 'r') as env_results_file_content:
         env_results = json.load(env_results_file_content)
-        instance_dns = env_results[0]['InstanceDNS']
+
+    # TODO: update this code after switching to parallel tests using multiple server for nightly build
+    instance_dns = env_results[0]['InstanceDNS']
+    role = env_results[0]['Role']
+    success_file_path = "./Tests/is_build_passed_{}.txt".format(role.replace(' ', ''))
+
+    content_team_fields, content_fields, _, _ = get_fields()
+    color = 'good' if os.path.isfile(success_file_path) else 'danger'
+    title = 'Content Build - Success' if os.path.isfile(success_file_path) else 'Content Build - Failure'
 
     content_team_attachment = [{
         'fallback': title,
         'color': color,
         'title': title,
         'title_link': build_url,
-        "author_name": "Demisto AWS Machine",
+        "author_name": "Demisto Machine (Click here to open the nightly server)",
         "author_link": "https://{0}".format(instance_dns),
         "author_icon": DEMISTO_GREY_ICON,
         'fields': content_team_fields
@@ -74,20 +76,33 @@ def get_attachments(build_url, env_results_file_name):
 
 
 def get_fields():
-    print('Extracting failed_tests')
-    with open('./Tests/failed_tests.txt', 'r') as failed_tests_file:
-        failed_tests = failed_tests_file.readlines()
-        failed_tests = [line.strip('\n') for line in failed_tests]
+    failed_tests = []
+    if os.path.isfile('./Tests/failed_tests.txt'):
+        print('Extracting failed_tests')
+        with open('./Tests/failed_tests.txt', 'r') as failed_tests_file:
+            failed_tests = failed_tests_file.readlines()
+            failed_tests = [line.strip('\n') for line in failed_tests]
 
-    print('Extracting skipped_tests')
-    with open('./Tests/skipped_tests.txt', 'r') as skipped_tests_file:
-        skipped_tests = skipped_tests_file.readlines()
-        skipped_tests = [line.strip('\n') for line in skipped_tests]
+    failed_unittests = []
+    if os.path.isfile('./Tests/failed_unittests.txt'):
+        print('Extracting failed_unittests')
+        with open('./Tests/failed_unittests.txt', 'r') as failed_unittests_file:
+            failed_unittests = failed_unittests_file.readlines()
+            failed_unittests = [line.strip('\n') for line in failed_unittests]
 
-    print('Extracting skipped_integrations')
-    with open('./Tests/skipped_integrations.txt', 'r') as skipped_integrations_file:
-        skipped_integrations = skipped_integrations_file.readlines()
-        skipped_integrations = [line.strip('\n') for line in skipped_integrations]
+    skipped_tests = []
+    if os.path.isfile('./Tests/skipped_tests.txt'):
+        print('Extracting skipped_tests')
+        with open('./Tests/skipped_tests.txt', 'r') as skipped_tests_file:
+            skipped_tests = skipped_tests_file.readlines()
+            skipped_tests = [line.strip('\n') for line in skipped_tests]
+
+    skipped_integrations = []
+    if os.path.isfile('./Tests/skipped_integrations.txt'):
+        print('Extracting skipped_integrations')
+        with open('./Tests/skipped_integrations.txt', 'r') as skipped_integrations_file:
+            skipped_integrations = skipped_integrations_file.readlines()
+            skipped_integrations = [line.strip('\n') for line in skipped_integrations]
 
     content_team_fields = []
     content_fields = []
@@ -100,6 +115,15 @@ def get_fields():
         }
         content_team_fields.append(field_failed_tests)
         content_fields.append(field_failed_tests)
+
+    if failed_unittests:
+        field_failed_unittests = {
+            "title": "Failed unittests - ({})".format(len(failed_unittests)),
+            "value": '\n'.join(failed_unittests),
+            "short": False
+        }
+        content_team_fields.append(field_failed_unittests)
+        content_fields.append(field_failed_unittests)
 
     if skipped_tests:
         field_skipped_tests = {
@@ -117,22 +141,22 @@ def get_fields():
         }
         content_team_fields.append(field_skipped_integrations)
 
-    return content_team_fields, content_fields, failed_tests
+    return content_team_fields, content_fields, failed_tests, failed_unittests
 
 
 def slack_notifier(build_url, slack_token, env_results_file_name):
     branches = run_command("git branch")
-    branch_name_reg = re.search("\* (.*)", branches)
+    branch_name_reg = re.search(r'\* (.*)', branches)
     branch_name = branch_name_reg.group(1)
 
     if branch_name == 'master':
         print_color("Starting Slack notifications about nightly build", LOG_COLORS.GREEN)
         print("Extracting build status")
-        content_team_attachments, content_attachments = get_attachments(build_url, env_results_file_name)
+        content_team_attachments, _ = get_attachments(build_url, env_results_file_name)
 
-        print("Sending Slack messages to #content and #content-team")
-        sc = SlackClient(slack_token)
-        sc.api_call(
+        print("Sending Slack messages to #content-team")
+        slack_client = SlackClient(slack_token)
+        slack_client.api_call(
             "chat.postMessage",
             channel="dmst-content-team",
             username="Content CircleCI",
@@ -141,13 +165,13 @@ def slack_notifier(build_url, slack_token, env_results_file_name):
         )
 
 
-if __name__ == "__main__":
+def main():
     options = options_handler()
     if options.nightly:
         slack_notifier(options.url, options.slack, options.env_results_file_name)
     else:
         print_color("Not nightly build, stopping Slack Notifications about Content build", LOG_COLORS.RED)
 
-    os.remove('./Tests/failed_tests.txt')
-    os.remove('./Tests/skipped_tests.txt')
-    os.remove('./Tests/skipped_integrations.txt')
+
+if __name__ == '__main__':
+    main()