+++ Integrations/Nmap/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+
+
+## [19.11.1] - 2019-11-26
 Fixed an issue in nmap scans with '-sn' flag.